Update random.move
In the seed function, the line sui move build; seems out of place and may cause a syntax error. It should be removed if it's not intended to be a part of the function.

In the bytes_to_u64 function, the return type should be specified as -> u64 after the function signature.

In the bytes_to_u64 function, the variables value and i should be declared as mutable using the mut keyword.

In the rand_u64_with_seed function, the return value is not explicitly specified using the return keyword. It should be return bytes_to_u64(_seed);.

In the rand_u64_range_with_seed function, the expression (value % (high - low)) + low should be enclosed in curly braces to make it explicit that the addition is being applied to the result of the modulo operation.

The rand_u64 and rand_u64_range functions are marked as public which is not a valid Rust keyword. Instead, the intended access level should be pub.

There are several missing semicolons at the end of some statements, such as in the rand_u64_range_with_seed function.